+++ src/config.js
@@ -11,7 +11,7 @@ class Config {
   }
 
   get tagVersion() {
-    return 'r11.3.4';
+    return 'r11.3.5';
   }
 
   get appName() {
@@ -196,6 +196,7 @@ class Config {
       modalCameraView: 'modalCameraView',
       modalPopup: 'modalPopup',
       modalInput: 'modalInput',
+      modalConfirm: 'modalConfirm',
 
       //  Reset password
       resetPassword: 'resetPassword',