+++ src/constants/cart/types.js
@@ -1,0 +1,4 @@
+export const CART_TYPES = {
+    NORMAL: 'cart',
+    DROP_SHIP: 'dropship'
+}